+++ instrumentation/faraday/lib/opentelemetry/instrumentation/faraday/middlewares/tracer_middleware.rb
@@ -25,11 +25,13 @@ module OpenTelemetry
 
           def call(env)
             http_method = HTTP_METHODS_SYMBOL_TO_STRING[env.method]
+            config = Faraday::Instrumentation.instance.config
+
             attributes = span_creation_attributes(
-              http_method: http_method, url: env.url
+              http_method: http_method, url: env.url, config: config
             )
             tracer.in_span(
-              "HTTP #{http_method}", attributes: attributes, kind: :client
+              "HTTP #{http_method}", attributes: attributes, kind: config.fetch(:span_kind)
             ) do |span|
               OpenTelemetry.propagation.inject(env.request_headers)
 
@@ -39,20 +41,22 @@ module OpenTelemetry
 
           private
 
-          attr_reader :app
-
-          def span_creation_attributes(http_method:, url:)
+          def span_creation_attributes(http_method:, url:, config:)
             instrumentation_attrs = {
               'http.method' => http_method,
               'http.url' => OpenTelemetry::Common::Utilities.cleanse_url(url.to_s),
-              'net.peer.name' => url.host
+              'faraday.adapter.name' => app.class.name
             }
-            config = Faraday::Instrumentation.instance.config
+            instrumentation_attrs['net.peer.name'] = url.host if url.host
             instrumentation_attrs['peer.service'] = config[:peer_service] if config[:peer_service]
+
             instrumentation_attrs.merge!(
               OpenTelemetry::Common::HTTP::ClientContext.attributes
             )
           end
+
+          # Versions prior to 1.0 do not define an accessor for app
+          attr_reader :app if Gem::Version.new(Faraday::VERSION) < Gem::Version.new('1.0.0')
 
           def tracer
             Faraday::Instrumentation.instance.tracer
@@ -60,7 +64,7 @@ module OpenTelemetry
 
           def trace_response(span, response)
             span.set_attribute('http.status_code', response.status)
-            span.status = OpenTelemetry::Trace::Status.error unless (100..399).include?(response.status.to_i)
+            span.status = OpenTelemetry::Trace::Status.error unless (100..399).cover?(response.status.to_i)
           end
         end
       end