+++ instrumentation/graphql/test/instrumentation/graphql/instrumentation_test.rb
@@ -54,7 +54,7 @@ describe OpenTelemetry::Instrumentation::GraphQL do
           OpenTelemetry::TestHelpers.with_test_logger do |log|
             instrumentation.install(config)
             _(log.string).must_match(
-              / Unable to patch schema Old::Truck: undefined method `trace_with' for Old::Truck:Class/
+              /undefined method `trace_with'.*Old::Truck/
             )
           end
         end
@@ -90,7 +90,7 @@ describe OpenTelemetry::Instrumentation::GraphQL do
             instrumentation.install(config)
 
             _(log.string).must_match(
-              /Unable to patch schema Old::Truck: undefined method `use' for Old::Truck:Class/
+              /undefined method `use'.*Old::Truck/
             )
           end
         end