+++ instrumentation/pg/test/opentelemetry/instrumentation/pg/instrumentation_test.rb
@@ -51,7 +51,7 @@ describe OpenTelemetry::Instrumentation::PG::Instrumentation do
     let(:user) { ENV.fetch('TEST_POSTGRES_USER', 'postgres') }
     let(:dbname) { ENV.fetch('TEST_POSTGRES_DB', 'postgres') }
     let(:password) { ENV.fetch('TEST_POSTGRES_PASSWORD', 'postgres') }
-
+    let(:config) { { db_statement: :include } }
     before do
       instrumentation.install(config)
     end
@@ -346,7 +346,7 @@ describe OpenTelemetry::Instrumentation::PG::Instrumentation do
         client.query('SELECT 1')
 
         _(span.attributes['net.peer.name']).must_equal host
-        _(span.attributes['net.peer.port']).must_equal port.to_i if PG.const_defined?('DEF_PORT')
+        _(span.attributes['net.peer.port']).must_equal port.to_i if PG.const_defined?(:DEF_PORT)
       end
     end
   end unless ENV['OMIT_SERVICES']