+++ instrumentation/rack/test/opentelemetry/instrumentation/rack/instrumentation_test.rb
@@ -13,7 +13,7 @@ describe OpenTelemetry::Instrumentation::Rack::Instrumentation do
 
   before do
     # simulate a fresh install:
-    instrumentation.instance_variable_set('@installed', false)
+    instrumentation.instance_variable_set(:@installed, false)
     instrumentation.config.clear
   end
 
@@ -32,7 +32,7 @@ describe OpenTelemetry::Instrumentation::Rack::Instrumentation do
       _(instrumentation.config[:url_quantization]).must_be_nil
       _(instrumentation.config[:untraced_requests]).must_be_nil
       _(instrumentation.config[:response_propagators]).must_be_empty
-      _(instrumentation.config[:use_rack_events]).must_equal false
+      _(instrumentation.config[:use_rack_events]).must_equal true
     end
   end