+++ instrumentation/action_view/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Release History: opentelemetry-instrumentation-action_view
 
+### v0.7.0 / 2023-11-22
+
+* BREAKING CHANGE: Drop Rails 6.0 EOL
+
+* ADDED: Drop Rails 6.0 EOL
+
+### v0.6.1 / 2023-10-16
+
+* FIXED: Add Rails 7.1 compatibility
+
 ### v0.6.0 / 2023-06-05
 
 * ADDED: Render layout for action view